Adding page

+++ scripts/build.sh
@@ -10,9 +10,20 @@ yarn --frozen-lockfile
 
 # # Clone non-modules
 function clone() {
-  rm -rf node_modules/$1 ; mkdir node_modules/$1
-  git clone $2 node_modules/$1
+  if [ ! -d node_modules/$1 ]
+  then
+    rm -rf node_modules/$1 ; mkdir node_modules/$1
+    git clone $2 node_modules/$1
+  fi
 }
+
+clone typeicons https://github.com/stephenhutchings/typicons.font.git ;
+
+clone githubicons https://github.com/primer/octicons.git ;
+
+clone gameicons https://github.com/delacannon/game-icons-inverted.git ;
+
+clone deviceicons https://github.com/vorillaz/devicons.git ;
 
 clone vscode https://github.com/microsoft/vscode-codicons.git ;
 
@@ -22,7 +33,9 @@ clone boxicons https://github.com/atisawd/boxicons.git ;
 
 clone icomoon https://github.com/Keyamoon/IcoMoon-Free.git ;
 
-clone ionicons https://github.com/ionic-team/ionicons.git 
+clone ionicons https://github.com/ionic-team/ionicons.git ;
+
+clone weathericons https://github.com/erikflowers/weather-icons.git ;
 
 # Build it
 node ./scripts/pack_creator.js
